+++ NEWS.md
@@ -1,6 +1,7 @@
 # h5lite 2.0.0.4
 
 * Added support for SZIP compression.
+* Added support for partial reading.